fix: 64-bit-correct function-pointer reads in cbase.h debug check
dlls/cbase.h FunctionCheck call sites in ThinkSet/TouchSet/UseSet/
BlockedSet read the entity's member function-pointer field via
  *((int*)((char*)this + offsetof(CBaseEntity, m_pfnX)))
which only reads 4 bytes. On 64-bit hosts member function pointers
occupy 8+ bytes, so the read picks up the lower half of the pointer
and silently feeds garbage to FunctionCheck. Replace `int*` with
`void**` to read the full pointer width on every arch. Debug-only
path (#ifdef _DEBUG); release builds don't compile this, so no
runtime behavior change in shipped MP binaries.

In FunctionCheck and CSave::WriteFunction, the cast
  NAME_FOR_FUNCTION((uint32)pFunction)
truncates a 64-bit pointer to 32 bits before passing to the engine
ABI (which itself takes uint32). Cast through uintptr_t to make the
narrowing explicit and silence -Wpointer-to-int-cast. Save/restore
remains broken on 64-bit until the engine widens pfnNameForFunction
/ pfnFunctionFromName to uintptr_t; this is documented in the
comments. Single-player saves only — MP doesn't exercise this path.

+++ dlls/util.cpp
@@ -1986,7 +1986,13 @@ void CSave::WriteFunction(const char* pname, void** data, int count)
 {
 	const char* functionName;
 
-	functionName = NAME_FOR_FUNCTION((uint32)*data);
+	// pfnNameForFunction's engine ABI takes uint32 — on 64-bit hosts the
+	// caller's full function pointer doesn't fit and silently truncates.
+	// Cast through uintptr_t to make the narrowing explicit and silence
+	// -Wpointer-to-int-cast. Save/restore via this path is therefore
+	// broken on 64-bit until the engine widens the API to uintptr_t;
+	// this only affects single-player saves (MP doesn't run save/restore).
+	functionName = NAME_FOR_FUNCTION((uint32)(uintptr_t)*data);
 	if (functionName)
 		BufferField(pname, strlen(functionName) + 1, functionName);
 	else
